+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `BaseContent` can now render any `children` blocks passed to it.
+- `minicart-checkout-button` interface.
+- New `minicartCheckoutButton` CSS handle.
+
+### Changed
+- Default blocks implementation to use these new features.
 
 ## [2.39.0] - 2020-01-22
 ### Added